refactor Android specific ChainCleaner code under AndroidPlatform
